chore(website): extract MobileTabBar constants to fix TS named-export errors [T001303]
Extract TABS and MOBILE_COL_INDEX from <script module> into mobile-tab-bar-constants.ts
so tsc can resolve named exports without the Svelte preprocessor. MobileTabBar.svelte
re-exports via const aliases to satisfy no-import-assign. Test now imports from the .ts file.
